news-mcp/requirements.txt
fastmcp
newsapi-python